perf(web): split releases display menu
